--- D3D11Window.cpp
@@ -1,180 +1,0 @@
-#include "D3D11Window.hpp"
-
-IMGUI_IMPL_API LRESULT ImGui_ImplWin32_WndProcHandler(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam);
-
-namespace FF7Remake 
-{
-	static uint64_t* MethodsTable;
-
-	D3D11Window::~D3D11Window() { }
-
-	LRESULT D3D11Window::WndProc(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam)
-	{
-		if (g_GameData->m_ShowMenu) 
-		{
-			ImGui_ImplWin32_WndProcHandler((HWND)g_D3D11Window->m_OldWndProc, msg, wParam, lParam);
-			return false;
-		}
-		return CallWindowProc((WNDPROC)g_D3D11Window->m_OldWndProc, hWnd, msg, wParam, lParam);
-	}
-
-	bool D3D11Window::GetWindowContext()
-	{
-		if (HijackWindow()) 
-		{
-			//	Update Pointers
-			g_Hooking->pSwapchain_Present = MethodsTable[8];
-			g_Hooking->pSwapchain_DrawIndexed = MethodsTable[12];
-			free(MethodsTable);
-			m_Init = true;
-			return true;
-		}
-		return false;
-	}
-
-	bool D3D11Window::HijackWindow()
-	{
-		if (!InitWindow())
-			return false;
-
-		HMODULE D3D11Module = GetModuleHandleA("d3d11.dll");
-		if (!D3D11Module)
-			return false;
-
-		D3D_FEATURE_LEVEL FeatureLevel;
-		const D3D_FEATURE_LEVEL FeatureLevels[] = { D3D_FEATURE_LEVEL_10_1, D3D_FEATURE_LEVEL_11_0 };
-
-		DXGI_RATIONAL RefreshRate;
-		RefreshRate.Numerator = 60;
-		RefreshRate.Denominator = 1;
-
-		DXGI_MODE_DESC BufferDesc;
-		BufferDesc.Width = 100;
-		BufferDesc.Height = 100;
-		BufferDesc.RefreshRate = RefreshRate;
-		BufferDesc.Format = DXGI_FORMAT_R8G8B8A8_UNORM;
-		BufferDesc.ScanlineOrdering = DXGI_MODE_SCANLINE_ORDER_UNSPECIFIED;
-		BufferDesc.Scaling = DXGI_MODE_SCALING_UNSPECIFIED;
-
-		DXGI_SAMPLE_DESC SampleDesc;
-		SampleDesc.Count = 1;
-		SampleDesc.Quality = 0;
-
-		DXGI_SWAP_CHAIN_DESC SwapChainDesc;
-		SwapChainDesc.BufferDesc = BufferDesc;
-		SwapChainDesc.SampleDesc = SampleDesc;
-		SwapChainDesc.BufferUsage = DXGI_USAGE_RENDER_TARGET_OUTPUT;
-		SwapChainDesc.BufferCount = 1;
-		SwapChainDesc.OutputWindow = WindowHwnd;
-		SwapChainDesc.Windowed = 1;
-		SwapChainDesc.SwapEffect = DXGI_SWAP_EFFECT_DISCARD;
-		SwapChainDesc.Flags = DXGI_SWAP_CHAIN_FLAG_ALLOW_MODE_SWITCH;
-
-		IDXGISwapChain* SwapChain;
-		ID3D11Device* Device;
-		ID3D11DeviceContext* Context;
-		if (D3D11CreateDeviceAndSwapChain(0, D3D_DRIVER_TYPE_HARDWARE, 0, 0, FeatureLevels, 1, D3D11_SDK_VERSION, &SwapChainDesc, &SwapChain, &Device, &FeatureLevel, &Context) < 0)
-		{
-			DeleteWindow();
-			return false;
-		}
-
-		//	copy virtual method pointers into a table
-		MethodsTable = (uint64_t*)::calloc(205, sizeof(uint64_t));
-		memcpy(MethodsTable, *(uint64_t**)SwapChain, 18 * sizeof(uint64_t));
-		memcpy(MethodsTable + 18, *(uint64_t**)Device, 43 * sizeof(uint64_t));
-		memcpy(MethodsTable + 18 + 43, *(uint64_t**)Context, 144 * sizeof(uint64_t));
-
-		SwapChain->Release();
-		SwapChain = 0;
-		Device->Release();
-		Device = 0;
-		Context->Release();
-		Context = 0;
-		DeleteWindow();
-		return true;
-	}
-
-	bool D3D11Window::InitWindow()
-	{
-		WindowClass.cbSize = sizeof(WNDCLASSEX);
-		WindowClass.style = CS_HREDRAW | CS_VREDRAW;
-		WindowClass.lpfnWndProc = DefWindowProc;
-		WindowClass.cbClsExtra = 0;
-		WindowClass.cbWndExtra = 0;
-		WindowClass.hInstance = GetModuleHandle(0);
-		WindowClass.hIcon = 0;
-		WindowClass.hCursor = 0;
-		WindowClass.hbrBackground = 0;
-		WindowClass.lpszMenuName = 0;
-		WindowClass.lpszClassName = L"MJ";
-		WindowClass.hIconSm = 0;
-		RegisterClassEx(&WindowClass);
-		WindowHwnd = CreateWindow(WindowClass.lpszClassName, L"DX11 Window", WS_OVERLAPPEDWINDOW, 0, 0, 100, 100, 0, 0, WindowClass.hInstance, 0);
-		if (!WindowHwnd) 
-			return false;
-		return true;
-	}
-
-	bool D3D11Window::DeleteWindow()
-	{
-		DestroyWindow(WindowHwnd);
-		UnregisterClass(WindowClass.lpszClassName, WindowClass.hInstance);
-		if (WindowHwnd)
-			return false;
-		return true;
-	}
-
-	bool D3D11Window::InitImGui(IDXGISwapChain* swapChain)
-	{
-		if (SUCCEEDED(swapChain->GetDevice(__uuidof(ID3D11Device), (void**)&m_Device))) {
-			ImGui::CreateContext();
-			ImGuiIO& io = ImGui::GetIO(); (void)io;
-			ImGui::GetIO().WantCaptureMouse;
-			ImGui::GetIO().WantTextInput;
-			ImGui::GetIO().WantCaptureKeyboard;
-			io.ConfigFlags = ImGuiConfigFlags_NavEnableGamepad;
-			io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;
-			io.IniFilename = NULL;
-			m_Device->GetImmediateContext(&m_DeviceContext);
-
-			DXGI_SWAP_CHAIN_DESC Desc;
-			swapChain->GetDesc(&Desc);
-			g_GameData->g_GameWindow = Desc.OutputWindow;
-
-			ID3D11Texture2D* BackBuffer;
-			swapChain->GetBuffer(0, __uuidof(ID3D11Texture2D), (LPVOID*)&BackBuffer);
-			m_Device->CreateRenderTargetView(BackBuffer, NULL, &m_RenderTargetView);
-			BackBuffer->Release();
-
-			ImGui_ImplWin32_Init(g_GameData->g_GameWindow);
-			ImGui_ImplDX11_Init(m_Device, m_DeviceContext);
-			ImGui_ImplDX11_CreateDeviceObjects();
-			ImGui::GetIO().ImeWindowHandle = g_GameData->g_GameWindow;
-			m_OldWndProc = (WNDPROC)SetWindowLongPtr(g_GameData->g_GameWindow, GWLP_WNDPROC, (__int3264)(LONG_PTR)WndProc);
-			b_ImGui_Initialized = true;
-			return true;
-		}
-		b_ImGui_Initialized = false;
-		return false;
-	}
-
-	void D3D11Window::Overlay(IDXGISwapChain* pSwapChain)
-	{
-		if (!b_ImGui_Initialized)
-			InitImGui(pSwapChain);
-
-		ImGui_ImplDX11_NewFrame();
-		ImGui_ImplWin32_NewFrame();
-		ImGui::NewFrame();
-		ImGui::GetIO().MouseDrawCursor = g_GameData->m_ShowMenu;
-
-		g_Menu->Draw();
-
-		ImGui::EndFrame();
-		ImGui::Render();
-		m_DeviceContext->OMSetRenderTargets(1, &m_RenderTargetView, 0);
-		ImGui_ImplDX11_RenderDrawData(ImGui::GetDrawData());
-	}
-
-}
