+++ include/Menu.h
@@ -1,20 +1,17 @@
 #pragma once
 #include "helper.h"
-#include "Game.hpp"
-#include "D3D11Window.hpp"
 namespace FF7Remake 
 {
 	class Menu
 	{
 	public:
-		void Draw();
-		void MainMenu();
-		void HUD();
-		void Loops();
+		static void Draw();
+		static void MainMenu();
+		static void HUD();
+		static void Loops();
 
 		//	constructor
 		Menu()  noexcept = default;
 		~Menu() noexcept = default;
 	};
-	inline std::unique_ptr<Menu> g_Menu;
 }
