+++ src/Game.cpp
@@ -1,133 +1,176 @@
-﻿#include "Hooking.hpp"
+#include "Game.h"
 namespace FF7Remake 
 {
-	Hooking::Hooking() { MH_Initialize(); }
+	//	STATICS
+	AGameBase*										AGame::gGameBase{ nullptr };
+	bool											AGame::bDemiGod{ false };
+	bool											AGame::bDemiGodMagic{ false };
+	bool											AGame::bMaxLimit{ false };
+	bool											AGame::bMaxATB{ false };
+	bool											AGame::bPauseGame{ false };
+	bool											AGame::bNullDmg{ false };
+	bool											AGame::bNullMgk{ false };
+	bool											AGame::bNullItem{ false };
+	bool											AGame::bModTimeScale{ false };
+	float											AGame::fTimeScale{ 1.0f };
+	bool											AGame::bMaxTargetLevel{ false };
+	bool											AGame::bKillTarget{ false };
+	bool											AGame::bNoTargetAttack{ false };		//	target doesnt attack
+	bool											AGame::bNullTargetDmg{ false };			//	target takes no damage
+	bool											AGame::bTargetAlwaysStagger{ false };	//	target defense is set to 0
+	bool											AGame::bXpFarn{ false };				//	sets targets hp to 0, prevents targets from attacking , sets target stagger to max and sets target to max level for max reward
+	__int64											AGame::Hooks::pXInput_State{ 0 };
+	__int64											AGame::Hooks::pAScene_Update{ 0 };
+	__int64											AGame::Hooks::pAPlayerState_SetHealth{ 0 };
+	__int64											AGame::Hooks::pAPlayerState_SetMana{ 0 };
+	__int64											AGame::Hooks::pAPlayerState_SubItem{ 0 };
+	__int64											AGame::Hooks::pATargetEntity_GetHP{ 0 };
+	__int64											AGame::Hooks::pATargetEntity_GetStaggerAmount{ 0 };
+	AGame::Hooks::XInput_State						AGame::Hooks::XInput_State_stub;
+	AGame::Hooks::AScene_Update						AGame::Hooks::AScene_Update_stub;
+	AGame::Hooks::APlayerState_SetHealth			AGame::Hooks::APlayerState_SetHealth_stub;
+	AGame::Hooks::APlayerState_SetMP				AGame::Hooks::APlayerState_SetMana_stub;
+	AGame::Hooks::APlayerState_SubItem				AGame::Hooks::APlayerState_SubItem_stub;
+	AGame::Hooks::ATargetEntity_GetHP				AGame::Hooks::ATargetEntity_GetHP_stub;
+	AGame::Hooks::ATargetEntity_GetStaggerAmount	AGame::Hooks::ATargetEntity_GetStaggerAmount_stub;
 
-	Hooking::~Hooking()
+	//
+	//	AGame
+	//
+#pragma region	//	AGAME
+
+	void AGame::InitGame()
 	{
-		DisableAllHooks();
-		RemoveAllHooks();
-		MH_Uninitialize();
-	}
 
-	void Hooking::Initialize()
-	{
-		//	Resolve Function Pointers
-		pXInput_State				= (g_GameData->g_GameBaseAddr + Offsets::oXinputState);
-		pAScene_Update				= (g_GameData->g_GameBaseAddr + Offsets::oSceneUpdate);
-		pAPlayerState_SetHealth		= (g_GameData->g_GameBaseAddr + Offsets::oSetHealth);
-		pAPlayerState_SetMana		= (g_GameData->g_GameBaseAddr + Offsets::oSetMana);
-		pAPlayerState_SubItem		= (g_GameData->g_GameBaseAddr + Offsets::oSubItem);
-		pATargetEntity_GetHP		= (g_GameData->g_GameBaseAddr + Offsets::oTargetGetHP);
-		pATargetEntity_GetStaggerAmount = (g_GameData->g_GameBaseAddr + Offsets::oTargetGetStaggerAmount);
+		gGameBase									= reinterpret_cast<AGameBase*>(g_Engine->g_GameBaseAddr + Offsets::oGameBase);
+		Hooks::pXInput_State						= (g_Engine->g_GameBaseAddr + Offsets::oXinputState);
+		Hooks::pAScene_Update						= (g_Engine->g_GameBaseAddr + Offsets::oSceneUpdate);
+		Hooks::pAPlayerState_SetHealth				= (g_Engine->g_GameBaseAddr + Offsets::oSetHealth);
+		Hooks::pAPlayerState_SetMana				= (g_Engine->g_GameBaseAddr + Offsets::oSetMana);
+		Hooks::pAPlayerState_SubItem				= (g_Engine->g_GameBaseAddr + Offsets::oSubItem);
+		Hooks::pATargetEntity_GetHP					= (g_Engine->g_GameBaseAddr + Offsets::oTargetGetHP);
+		Hooks::pATargetEntity_GetStaggerAmount		= (g_Engine->g_GameBaseAddr + Offsets::oTargetGetStaggerAmount);
 
-		if (Console::m_bGUI && g_D3D11Window->GetWindowContext())
-		{
-			if (pSwapchain_Present)
-				CreateHook((LPVOID)pSwapchain_Present, &Swapchain_present_hook, (void**)&Swapchain_present_stub);
 
-			if (pSwapchain_DrawIndexed)
-				CreateHook((LPVOID)pSwapchain_DrawIndexed, &Swapchain_DrawIndexed_hook, (void**)&Swapchain_DrawIndexed_stub);
-
-			g_GameData->m_ShowHud = true;
-		}
 
 		//	Controller Input Handler
-		if (pXInput_State)
-			CreateHook((LPVOID)pXInput_State, &XInput_State_hook, (void**)&XInput_State_stub);
+		if (Hooks::pXInput_State)
+			Hooking::CreateHook((LPVOID)Hooks::pXInput_State, &Hooks::XInput_State_hook, (void**)&Hooks::XInput_State_stub);
 
 		//	Render Scene Update 
-		if (pAScene_Update)
-			CreateHook((LPVOID)pAScene_Update, &AScene_Update_hook, (void**)&AScene_Update_stub);
+		if (Hooks::pAScene_Update)
+			Hooking::CreateHook((LPVOID)Hooks::pAScene_Update, &Hooks::AScene_Update_hook, (void**)&Hooks::AScene_Update_stub);
 
 		//	Party Set Health Handler
-		if (pAPlayerState_SetHealth)
-			CreateHook((LPVOID)pAPlayerState_SetHealth, &APlayerState_SetHealth_hook, (void**)&APlayerState_SetHealth_stub);
-		
+		if (Hooks::pAPlayerState_SetHealth)
+			Hooking::CreateHook((LPVOID)Hooks::pAPlayerState_SetHealth, &Hooks::APlayerState_SetHealth_hook, (void**)&Hooks::APlayerState_SetHealth_stub);
+
 		//	Party Set Mana Handler
-		if (pAPlayerState_SetMana)
-			CreateHook((LPVOID)pAPlayerState_SetMana, &APlayerState_SetMana_hook, (void**)&APlayerState_SetMana_stub);
+		if (Hooks::pAPlayerState_SetMana)
+			Hooking::CreateHook((LPVOID)Hooks::pAPlayerState_SetMana, &Hooks::APlayerState_SetMana_hook, (void**)&Hooks::APlayerState_SetMana_stub);
 
 		//	Party Sub Item Handler
-		if (pAPlayerState_SubItem)
-			CreateHook((LPVOID)pAPlayerState_SubItem, &APlayerState_SubItem_hook, (void**)&APlayerState_SubItem_stub);
+		if (Hooks::pAPlayerState_SubItem)
+			Hooking::CreateHook((LPVOID)Hooks::pAPlayerState_SubItem, &Hooks::APlayerState_SubItem_hook, (void**)&Hooks::APlayerState_SubItem_stub);
 
 		//	Target Entity Handler
-		if (pATargetEntity_GetHP)
-			CreateHook((LPVOID)pATargetEntity_GetHP, &ATargetEntity_GetHP_hook, (void**)&ATargetEntity_GetHP_stub);
+		if (Hooks::pATargetEntity_GetHP)
+			Hooking::CreateHook((LPVOID)Hooks::pATargetEntity_GetHP, &Hooks::ATargetEntity_GetHP_hook, (void**)&Hooks::ATargetEntity_GetHP_stub);
 
 		//	Target Entity Stagger Handler
-		if (pATargetEntity_GetStaggerAmount)
-			CreateHook((LPVOID)pATargetEntity_GetStaggerAmount, &ATargetEntity_GetStaggerAmount_hook, (void**)&ATargetEntity_GetStaggerAmount_stub);
-
-		EnableAllHooks();
+		if (Hooks::pATargetEntity_GetStaggerAmount)
+			Hooking::CreateHook((LPVOID)Hooks::pATargetEntity_GetStaggerAmount, &Hooks::ATargetEntity_GetStaggerAmount_hook, (void**)&Hooks::ATargetEntity_GetStaggerAmount_stub);
 	}
 
-	void Hooking::Shutdown()
+	void AGame::ShutdownGame()
 	{
-		if (pSwapchain_Present)
-			DisableHook((LPVOID)pSwapchain_Present);
+		if (Hooks::pXInput_State)
+			Hooking::DisableHook((LPVOID)Hooks::pXInput_State);
 
-		if (pSwapchain_DrawIndexed)
-			DisableHook((LPVOID)pSwapchain_DrawIndexed);
+		if (Hooks::pAScene_Update)
+			Hooking::DisableHook((LPVOID)Hooks::pAScene_Update);
 
-		if (pXInput_State)
-			DisableHook((LPVOID)pXInput_State);
-		
-		if (pAScene_Update)
-			DisableHook((LPVOID)pAScene_Update);
+		if (Hooks::pAPlayerState_SetHealth)
+			Hooking::DisableHook((LPVOID)Hooks::pAPlayerState_SetHealth);
 
-		if (pAPlayerState_SetHealth)
-			DisableHook((LPVOID)pAPlayerState_SetHealth);
+		if (Hooks::pAPlayerState_SetMana)
+			Hooking::DisableHook((LPVOID)Hooks::pAPlayerState_SetMana);
 
-		if (pAPlayerState_SetMana)
-			DisableHook((LPVOID)pAPlayerState_SetMana);
+		if (Hooks::pAPlayerState_SubItem)
+			Hooking::DisableHook((LPVOID)Hooks::pAPlayerState_SubItem);
 
-		if (pAPlayerState_SubItem)
-			DisableHook((LPVOID)pAPlayerState_SubItem);
+		if (Hooks::pATargetEntity_GetHP)
+			Hooking::DisableHook((LPVOID)Hooks::pATargetEntity_GetHP);
 
-		if (pATargetEntity_GetHP)
-			DisableHook((LPVOID)pATargetEntity_GetHP);
-
-		if (pATargetEntity_GetStaggerAmount)
-			DisableHook((LPVOID)pATargetEntity_GetStaggerAmount);
-
-		RemoveAllHooks();
+		if (Hooks::pATargetEntity_GetStaggerAmount)
+			Hooking::DisableHook((LPVOID)Hooks::pATargetEntity_GetStaggerAmount);
 	}
 
-	bool Hooking::CreateHook(LPVOID lpTarget, LPVOID pDetour, LPVOID* pOrig)
+	void AGame::Patches::RefillCloudHP()
 	{
-		if (MH_CreateHook(lpTarget, pDetour, pOrig) != MH_OK || MH_EnableHook(lpTarget) != MH_OK)
-			return false;
-		return true;
+		AGameState* pGameState = gGameBase->GetGameState();
+		if (!pGameState)
+			return;
+
+		//	Get cloud player stats
+		APlayerStats cloud_stats = pGameState->GetCloudStats();
+		cloud_stats.RefillHP();	//	set hp var to match max hp
+
+		//	 Set cloud new player stats
+		pGameState->SetCloudStats(cloud_stats);
 	}
 
-	void Hooking::EnableHook(LPVOID lpTarget) { MH_EnableHook(lpTarget); }
-
-	void Hooking::DisableHook(LPVOID lpTarget) { MH_DisableHook(lpTarget); }
-
-	void Hooking::RemoveHook(LPVOID lpTarget) { MH_RemoveHook(lpTarget); }
-
-	void Hooking::EnableAllHooks() { MH_EnableHook(MH_ALL_HOOKS); }
-
-	void Hooking::DisableAllHooks() { MH_DisableHook(MH_ALL_HOOKS); }
-
-	void Hooking::RemoveAllHooks() { MH_RemoveHook(MH_ALL_HOOKS); }
-
-	HRESULT APIENTRY Hooking::Swapchain_present_hook(IDXGISwapChain* pSwapChain, UINT SyncInterval, UINT Flags)
-	{	
-		//	Render Scene
-		g_D3D11Window->Overlay(pSwapChain);
-		
-		return g_Hooking->Swapchain_present_stub(pSwapChain, SyncInterval, Flags);
-	}
-
-	void APIENTRY Hooking::Swapchain_DrawIndexed_hook(ID3D11DeviceContext* pContext, UINT IndexCount, UINT StartIndexLocation, INT BaseVertexLocation)
+	void AGame::Patches::RefillCloudMP()
 	{
-		g_Hooking->Swapchain_DrawIndexed_stub(pContext, IndexCount, StartIndexLocation, BaseVertexLocation);
+		AGameState* pGameState = gGameBase->GetGameState();
+		if (!pGameState)
+			return;
+
+		//	Get cloud player stats
+		APlayerStats cloud_stats = pGameState->GetCloudStats();
+		cloud_stats.RefillMana();
+
+		//	 Set cloud new player stats
+		pGameState->SetCloudStats(cloud_stats);
 	}
 
-	__int64 Hooking::XInput_State_hook(__int64 a1)
+	void AGame::Patches::CloudMaxLimit()
+	{
+		AGameState* pGameState = gGameBase->GetGameState();
+		if (!pGameState)
+			return;
+
+		//	Get cloud player stats
+		APlayerStats cloud_stats = pGameState->GetCloudStats();
+		cloud_stats.SetMaxLimit();
+
+		//	 Set cloud new player stats
+		pGameState->SetCloudStats(cloud_stats);
+	}
+
+	void AGame::Patches::CloudMaxATB()
+	{
+		AGameState* pGameState = gGameBase->GetGameState();
+		if (!pGameState)
+			return;
+
+		//	Get cloud player stats
+		APlayerStats cloud_stats = pGameState->GetCloudStats();
+		cloud_stats.SetMaxATB();
+
+		//	 Set cloud new player stats
+		pGameState->SetCloudStats(cloud_stats);
+	}
+
+#pragma endregion
+
+	//
+	//	AGame::Hooks
+	//
+#pragma region	//	AGAME::HOOKS
+
+	//	
+	__int64 AGame::Hooks::XInput_State_hook(__int64 a1)
 	{
 		///	PORT NOTES
 		/*
@@ -135,13 +178,14 @@ namespace FF7Remake
 		*/
 
 		//	Whenever the menu is shown , we wont pass controller inputs to the game
-		if (g_GameData->m_ShowMenu)
+		if (g_Engine->m_ShowMenu)
 			return 0;
 
-		g_Hooking->XInput_State_stub(a1);
+		XInput_State_stub(a1);
 	}
 
-	void Hooking::AScene_Update_hook(__int64 p)
+	//	
+	void AGame::Hooks::AScene_Update_hook(__int64 p)
 	{
 		/// PORT NOTES
 		/*
@@ -169,24 +213,24 @@ namespace FF7Remake
 		*/
 
 		auto pScene = reinterpret_cast<AScene*>(p);
-		if (pScene && g_GameData->m_ShowMenu)
+		if (pScene && g_Engine->m_ShowMenu)
 		{
 			//	pScene->SetPauseState(true);
 			pScene->SetTimeScale(0.00f);
 		}
 
-		if (pScene && g_GameData->bPauseGame)
+		if (pScene && AGame::bPauseGame)
 			pScene->SetTimeScale(pScene->GetTimeScale() * 0.0f);
 
-		if (pScene && g_GameData->bModTimeScale)
-			pScene->SetTimeScale(g_GameData->fTimeScale);
+		if (pScene && AGame::bModTimeScale)
+			pScene->SetTimeScale(AGame::fTimeScale);
 
 		//	exec original fn
-		g_Hooking->AScene_Update_stub(p);
+		AScene_Update_stub(p);
 	}
 
 	//	sets hp to party members by traditional game mechanics
-	__int64 Hooking::APlayerState_SetHealth_hook(unsigned __int8 index, int health)
+	__int64 AGame::Hooks::APlayerState_SetHealth_hook(unsigned __int8 index, int health)
 	{
 		///	PORT NOTES
 		/*
@@ -216,7 +260,7 @@ namespace FF7Remake
 			.text:0000000140AFB6E7                 call    sub_141C22CC0																//
 			.text:0000000140AFB6EC                 mov     rcx, cs:qword_1457F5148														//
 			.text:0000000140AFB6F3																										//
-			.text:0000000140AFB6F3 loc_140AFB6F3:                          ; CODE XREF: APlayerState_SubHealth+25↑j						//
+			.text:0000000140AFB6F3 loc_140AFB6F3:                          																//
 			.text:0000000140AFB6F3                 mov     rax, [rcx]																	//
 			.text:0000000140AFB6F6                 xor     r8d, r8d																		//
 			.text:0000000140AFB6F9                 mov     edx, 0B6DF0h																	//
@@ -226,10 +270,10 @@ namespace FF7Remake
 			.text:0000000140AFB706                 mov     rcx, rax																		//
 			.text:0000000140AFB709                 call    sub_140B26300																//
 			.text:0000000140AFB70E																										//
-			.text:0000000140AFB70E loc_140AFB70E:                          ; CODE XREF: APlayerState_SubHealth+44↑j						//
+			.text:0000000140AFB70E loc_140AFB70E:                          																//
 			.text:0000000140AFB70E                 mov     cs:qword_1457B9268, rax														//
 			.text:0000000140AFB715																										//
-			.text:0000000140AFB715 loc_140AFB715:                          ; CODE XREF: APlayerState_SubHealth+19↑j						//
+			.text:0000000140AFB715 loc_140AFB715:                          																//
 			.text:0000000140AFB715                 mov     rax, [rax+0B6D80h]															//
 			.text:0000000140AFB71C                 cmp     bl, 8																		//
 			.text:0000000140AFB71F                 jnb     short loc_140AFB73C															//
@@ -240,8 +284,7 @@ namespace FF7Remake
 			.text:0000000140AFB734                 shl     rcx, 6																		//
 			.text:0000000140AFB738                 mov     [rcx+rax+30h], edi															//	[!] This instruction subtracts hp from party members
 			.text:0000000140AFB73C																										//
-			.text:0000000140AFB73C loc_140AFB73C:                          ; CODE XREF: APlayerState_SubHealth+5F↑j						//
-			.text:0000000140AFB73C                                         ; APlayerState_SubHealth+68↑j								//
+			.text:0000000140AFB73C loc_140AFB73C:                          																//
 			.text:0000000140AFB73C                 mov     rbx, [rsp+28h+arg_0]															//
 			.text:0000000140AFB741                 add     rsp, 20h																		//
 			.text:0000000140AFB745                 pop     rdi																			//
@@ -253,19 +296,19 @@ namespace FF7Remake
 		auto pSetHealth = OnSetHealth(index, health);
 
 		if (g_Console->m_bVerbose && pSetHealth.mDiff > 0)
-			Console::Log("[+] [Hooking::APlayerState_SetHealth_hook(%d, %d)]\n- diff: %d\n- isTakeDmg: %d\n- isHealing: %d\n- origHealth: %d / %d\n- newHealth: %d / %d\n- bNullDmg: %d\n\n", 
-				index, health, pSetHealth.mDiff, pSetHealth.bTakingDmg, pSetHealth.bHealing, health + pSetHealth.mDiff, pSetHealth.mPlayerStats.MaxHP, health, pSetHealth.mPlayerStats.MaxHP, g_GameData->bNullDmg);
-		
+			Console::Log("[+] [Hooking::APlayerState_SetHealth_hook(%d, %d)]\n- diff: %d\n- isTakeDmg: %d\n- isHealing: %d\n- origHealth: %d / %d\n- newHealth: %d / %d\n- bNullDmg: %d\n\n",
+				index, health, pSetHealth.mDiff, pSetHealth.bTakingDmg, pSetHealth.bHealing, health + pSetHealth.mDiff, pSetHealth.mPlayerStats.MaxHP, health, pSetHealth.mPlayerStats.MaxHP, AGame::bNullDmg);
+
 		//	prevent damage to party member
-		if (g_GameData->bNullDmg && pSetHealth.bTakingDmg)
+		if (AGame::bNullDmg && pSetHealth.bTakingDmg)
 			return 0;
 
 		//	exec original fn
-		return g_Hooking->APlayerState_SetHealth_stub(index, health);
+		return APlayerState_SetHealth_stub(index, health);
 	}
 
 	//	sets mp to party members by traditional game mechanics
-	__int64 Hooking::APlayerState_SetMana_hook(unsigned __int8 index, int mana)
+	__int64 AGame::Hooks::APlayerState_SetMana_hook(unsigned __int8 index, int mana)
 	{
 		///	PORT NOTES
 		/*
@@ -276,8 +319,7 @@ namespace FF7Remake
 
 
 			.text:0000000140AFB8D0 ; __int64 __fastcall APlayerState_SetMp(unsigned __int8, int)
-			.text:0000000140AFB8D0 APlayerState_SetMp proc near            ; CODE XREF: sub_1409498C0+4↑j
-			.text:0000000140AFB8D0                                         ; sub_1409A1BA0+ED↑p ...
+			.text:0000000140AFB8D0 APlayerState_SetMp proc near
 			.text:0000000140AFB8D0
 			.text:0000000140AFB8D0 arg_0           = qword ptr  8
 			.text:0000000140AFB8D0
@@ -295,7 +337,7 @@ namespace FF7Remake
 			.text:0000000140AFB8F7                 call    sub_141C22CC0
 			.text:0000000140AFB8FC                 mov     rcx, cs:qword_1457F5148
 			.text:0000000140AFB903
-			.text:0000000140AFB903 loc_140AFB903:                          ; CODE XREF: APlayerState_SetMp+25↑j
+			.text:0000000140AFB903 loc_140AFB903:
 			.text:0000000140AFB903                 mov     rax, [rcx]
 			.text:0000000140AFB906                 xor     r8d, r8d
 			.text:0000000140AFB909                 mov     edx, 0B6DF0h
@@ -305,10 +347,10 @@ namespace FF7Remake
 			.text:0000000140AFB916                 mov     rcx, rax
 			.text:0000000140AFB919                 call    sub_140B26300
 			.text:0000000140AFB91E
-			.text:0000000140AFB91E loc_140AFB91E:                          ; CODE XREF: APlayerState_SetMp+44↑j
+			.text:0000000140AFB91E loc_140AFB91E:
 			.text:0000000140AFB91E                 mov     cs:qword_1457B9268, rax
 			.text:0000000140AFB925
-			.text:0000000140AFB925 loc_140AFB925:                          ; CODE XREF: APlayerState_SetMp+19↑j
+			.text:0000000140AFB925 loc_140AFB925:
 			.text:0000000140AFB925                 cmp     bl, 8
 			.text:0000000140AFB928                 jnb     short loc_140AFB943
 			.text:0000000140AFB92A                 mov     rax, [rax+0B6D80h]
@@ -317,7 +359,7 @@ namespace FF7Remake
 			.text:0000000140AFB938                 mov     rax, [rax+3E60h]
 			.text:0000000140AFB93F                 mov     [rax+rcx+38h], edi
 			.text:0000000140AFB943
-			.text:0000000140AFB943 loc_140AFB943:                          ; CODE XREF: APlayerState_SetMp+58↑j
+			.text:0000000140AFB943 loc_140AFB943:
 			.text:0000000140AFB943                 mov     rbx, [rsp+28h+arg_0]
 			.text:0000000140AFB948                 add     rsp, 20h
 			.text:0000000140AFB94C                 pop     rdi
@@ -328,7 +370,7 @@ namespace FF7Remake
 			{
 			  __int64 result; // rax
 			  __int64 v5; // rcx
-			
+
 			  result = qword_1457B9268;
 			  if ( !qword_1457B9268 )
 			  {
@@ -357,18 +399,19 @@ namespace FF7Remake
 		auto pSetMana = OnSetMana(index, mana);
 
 		if (g_Console->m_bVerbose && pSetMana.mDiff > 0)
-			Console::Log("[+] [Hooking::APlayerState_SetMana_hook(%d, %d)]\n- diff: %d\n- isUsingMana: %d\n- isHealingMana: %d\n- origMana: %d / %d\n- newMana: %d / %d\n- bNullMgk: %d\n\n", 
-				index, mana, pSetMana.mDiff, pSetMana.bUsingMana, pSetMana.bHealing, mana + pSetMana.mDiff, pSetMana.mPlayerStats.MaxMP, mana, pSetMana.mPlayerStats.MaxMP, g_GameData->bNullMgk);
+			Console::Log("[+] [Hooking::APlayerState_SetMana_hook(%d, %d)]\n- diff: %d\n- isUsingMana: %d\n- isHealingMana: %d\n- origMana: %d / %d\n- newMana: %d / %d\n- bNullMgk: %d\n\n",
+				index, mana, pSetMana.mDiff, pSetMana.bUsingMana, pSetMana.bHealing, mana + pSetMana.mDiff, pSetMana.mPlayerStats.MaxMP, mana, pSetMana.mPlayerStats.MaxMP, AGame::bNullMgk);
 
 		//	prevent damage to party member
-		if (g_GameData->bNullMgk && pSetMana.bUsingMana)
+		if (AGame::bNullMgk && pSetMana.bUsingMana)
 			return 0;
 
 		//	exec original fn
-		return g_Hooking->APlayerState_SetMana_stub(index, mana);
+		return APlayerState_SetMana_stub(index, mana);
 	}
 
-	__int64 Hooking::APlayerState_SubItem_hook(__int64 p, int index)
+	//	
+	__int64 AGame::Hooks::APlayerState_SubItem_hook(__int64 p, int index)
 	{
 		///	PORT NOTES
 		/*
@@ -395,8 +438,7 @@ namespace FF7Remake
 			Figure out which function is adjusting the count value by checking what changes the item count value.
 
 			.text:0000000140B1CAF0 ; __int64 __fastcall APlayerState_SubItem(__int64, int)
-			.text:0000000140B1CAF0 APlayerState_SubItem proc near          ; CODE XREF: sub_140B020A0+F0↑p
-			.text:0000000140B1CAF0                                         ; DATA XREF: .rdata:0000000144FCFB14↓o ...
+			.text:0000000140B1CAF0 APlayerState_SubItem proc near
 			.text:0000000140B1CAF0
 			.text:0000000140B1CAF0 arg_0           = qword ptr  8
 			.text:0000000140B1CAF0 arg_8           = qword ptr  10h
@@ -415,8 +457,7 @@ namespace FF7Remake
 			.text:0000000140B1CB14                 test    rdi, rdi
 			.text:0000000140B1CB17                 jz      loc_140B1CC2A
 			.text:0000000140B1CB1D
-			.text:0000000140B1CB1D loc_140B1CB1D:                          ; DATA XREF: .rdata:0000000144FCFB14↓o
-			.text:0000000140B1CB1D                                         ; .rdata:0000000144FCFB24↓o ...
+			.text:0000000140B1CB1D loc_140B1CB1D:
 			.text:0000000140B1CB1D                 mov     [rsp+28h+arg_0], rbx
 			.text:0000000140B1CB22                 mov     [rsp+28h+arg_8], r14
 			.text:0000000140B1CB27                 call    sub_140FCB3D0
@@ -430,7 +471,7 @@ namespace FF7Remake
 			.text:0000000140B1CB4A                 mov     ecx, r14d
 			.text:0000000140B1CB4D                 nop     dword ptr [rax]
 			.text:0000000140B1CB50
-			.text:0000000140B1CB50 loc_140B1CB50:                          ; CODE XREF: APlayerState_SubItem+72↓j
+			.text:0000000140B1CB50 loc_140B1CB50:
 			.text:0000000140B1CB50                 cmp     [rax], r14
 			.text:0000000140B1CB53                 jz      short loc_140B1CB66
 			.text:0000000140B1CB55                 inc     ebx
@@ -441,14 +482,14 @@ namespace FF7Remake
 			.text:0000000140B1CB64                 jmp     short loc_140B1CBCC
 			.text:0000000140B1CB66 ; ---------------------------------------------------------------------------
 			.text:0000000140B1CB66
-			.text:0000000140B1CB66 loc_140B1CB66:                          ; CODE XREF: APlayerState_SubItem+63↑j
+			.text:0000000140B1CB66 loc_140B1CB66:
 			.text:0000000140B1CB66                 mov     rcx, cs:qword_1457F5148
 			.text:0000000140B1CB6D                 test    rcx, rcx
 			.text:0000000140B1CB70                 jnz     short loc_140B1CB7E
 			.text:0000000140B1CB72                 call    sub_141C22CC0
 			.text:0000000140B1CB77                 mov     rcx, cs:qword_1457F5148
 			.text:0000000140B1CB7E
-			.text:0000000140B1CB7E loc_140B1CB7E:                          ; CODE XREF: APlayerState_SubItem+80↑j
+			.text:0000000140B1CB7E loc_140B1CB7E:
 			.text:0000000140B1CB7E                 mov     rax, [rcx]
 			.text:0000000140B1CB81                 xor     r8d, r8d
 			.text:0000000140B1CB84                 lea     edx, [r8+18h]
@@ -463,10 +504,10 @@ namespace FF7Remake
 			.text:0000000140B1CBA5                 jmp     short loc_140B1CBAA
 			.text:0000000140B1CBA7 ; ---------------------------------------------------------------------------
 			.text:0000000140B1CBA7
-			.text:0000000140B1CBA7 loc_140B1CBA7:                          ; CODE XREF: APlayerState_SubItem+A1↑j
+			.text:0000000140B1CBA7 loc_140B1CBA7:
 			.text:0000000140B1CBA7                 mov     rcx, r14
 			.text:0000000140B1CBAA
-			.text:0000000140B1CBAA loc_140B1CBAA:                          ; CODE XREF: APlayerState_SubItem+B5↑j
+			.text:0000000140B1CBAA loc_140B1CBAA:
 			.text:0000000140B1CBAA                 movsxd  rax, ebx
 			.text:0000000140B1CBAD                 mov     [rbp+rax*8+368h], rcx
 			.text:0000000140B1CBB5                 test    rcx, rcx
@@ -478,8 +519,8 @@ namespace FF7Remake
 			.text:0000000140B1CBC6                 mov     eax, [rdi+0Ch]
 			.text:0000000140B1CBC9                 mov     [rcx+0Ch], eax
 			.text:0000000140B1CBCC
-			.text:0000000140B1CBCC loc_140B1CBCC:                          ; CODE XREF: APlayerState_SubItem+41↑j
-			.text:0000000140B1CBCC                                         ; APlayerState_SubItem+4A↑j ...
+			.text:0000000140B1CBCC loc_140B1CBCC:
+			.text:0000000140B1CBCC
 			.text:0000000140B1CBCC                 mov     rbx, [rsi+48h]
 			.text:0000000140B1CBD0                 mov     rdx, [rbx]
 			.text:0000000140B1CBD3                 test    rdx, rdx
@@ -493,7 +534,7 @@ namespace FF7Remake
 			.text:0000000140B1CBEE                 mov     [rbx+0Ch], r14d
 			.text:0000000140B1CBF2                 mov     rbx, [rsi+48h]
 			.text:0000000140B1CBF6
-			.text:0000000140B1CBF6 loc_140B1CBF6:                          ; CODE XREF: APlayerState_SubItem+E6↑j
+			.text:0000000140B1CBF6 loc_140B1CBF6:
 			.text:0000000140B1CBF6                 test    rbx, rbx
 			.text:0000000140B1CBF9                 jz      short loc_140B1CC20
 			.text:0000000140B1CBFB                 mov     rcx, cs:qword_1457F5148
@@ -502,18 +543,18 @@ namespace FF7Remake
 			.text:0000000140B1CC07                 call    sub_141C22CC0
 			.text:0000000140B1CC0C                 mov     rcx, cs:qword_1457F5148
 			.text:0000000140B1CC13
-			.text:0000000140B1CC13 loc_140B1CC13:                          ; CODE XREF: APlayerState_SubItem+115↑j
+			.text:0000000140B1CC13 loc_140B1CC13:
 			.text:0000000140B1CC13                 mov     rax, [rcx]
 			.text:0000000140B1CC16                 mov     rdx, rbx
 			.text:0000000140B1CC19                 call    qword ptr [rax+20h]
 			.text:0000000140B1CC1C                 mov     [rsi+48h], r14
 			.text:0000000140B1CC20
-			.text:0000000140B1CC20 loc_140B1CC20:                          ; CODE XREF: APlayerState_SubItem+109↑j
+			.text:0000000140B1CC20 loc_140B1CC20:
 			.text:0000000140B1CC20                 mov     rbx, [rsp+28h+arg_0]
 			.text:0000000140B1CC25                 mov     r14, [rsp+28h+arg_8]
 			.text:0000000140B1CC2A
-			.text:0000000140B1CC2A loc_140B1CC2A:                          ; CODE XREF: APlayerState_SubItem+27↑j
-			.text:0000000140B1CC2A                                         ; DATA XREF: .pdata:0000000145A771BC↓o ...
+			.text:0000000140B1CC2A loc_140B1CC2A:
+			.text:0000000140B1CC2A
 			.text:0000000140B1CC2A                 mov     rbp, [rsp+28h+arg_10]
 			.text:0000000140B1CC2F                 mov     rsi, [rsp+28h+arg_18]
 			.text:0000000140B1CC34                 add     rsp, 20h
@@ -529,7 +570,7 @@ namespace FF7Remake
 			v11 = *(__int64 **)(v4 + 0x48); // inventory
 			v12 = *v11;	// Slot
 			v13 = *(_DWORD *)(v12 + 0xC) - *((_DWORD *)v11 + 3); // sub item
-			
+
 			//	ADJUSTED FOR READABILITY
 			auto v4 = p + 8 * index;									// OnSubItem*
 			AInventory* v11 = *(__int64**)(v4 + 0x48);					// inventory
@@ -537,18 +578,18 @@ namespace FF7Remake
 			int sub_amount = *((__int32*)v11 + 3);						// sub amount
 			int v13 = v12->count - sub_amount;							// sub item new amount
 		*/
-		
+
 		///	
-		/*		
+		/*
 			///	RECREATION
 			//	auto pUseItem = reinterpret_cast<OnUseItem*> (p);
 			//	AItemSlot oItem = pUseItem->GetInventory()->Item;
-			
+
 			///	Get Other Slots
 			//	AItemSlot* slot = pUseItem->GetItemSlot(index + 1);
 			//	auto slot_item_count = slot->count;
 		*/
-		
+
 		auto v4 = p + 8 * index;									// OnSubItem*
 		auto v11 = *(__int64**)(v4 + 0x48);							// inventory
 		AItemSlot* v12 = reinterpret_cast<AItemSlot*>(*v11);		// ItemSlot
@@ -556,12 +597,12 @@ namespace FF7Remake
 		int v13 = v12->count - sub_amount;							// sub item new amount
 
 		if (g_Console->m_bVerbose)
-			Console::Log("[+] [Hooking::APlayerState_SubItem(0x%llX, %d)]\n- subAmount: %d\n- origCount: %d\n- newCount: %d\n- bNullItem: %d\n\n", 
-				p, index, sub_amount, v12->count, v13, g_GameData->bNullItem);
+			Console::Log("[+] [Hooking::APlayerState_SubItem(0x%llX, %d)]\n- subAmount: %d\n- origCount: %d\n- newCount: %d\n- bNullItem: %d\n\n",
+				p, index, sub_amount, v12->count, v13, AGame::bNullItem);
 
-		if (g_GameData->bNullItem)
+		if (AGame::bNullItem)
 		{
-			auto result = g_Hooking->APlayerState_SubItem_stub(p, index);
+			auto result = APlayerState_SubItem_stub(p, index);
 
 			//	Set new amount ( revert change )
 			v12->count = (v13 + sub_amount);
@@ -570,77 +611,11 @@ namespace FF7Remake
 		}
 
 		//	exec original fn
-		return g_Hooking->APlayerState_SubItem_stub(p, index);
+		return APlayerState_SubItem_stub(p, index);
 	}
 
-	struct ATarget
-	{
-		int		Level;			//0x0008
-		float	SpAtkTime;		//0x000C
-		float	SpAtkTimeMax;	//0x0010
-		float	AttackRate;		//0x0014
-		int		HP;				//0x0018
-		int		HPMax;			//0x001C
-		char	pad_0020[8];	//0x0020
-		int		Attack;			//0x0028
-		int		MagicAtk;		//0x002C
-		int		Defense;		//0x0030
-		int		MagicDefense;	//0x0034
-		char	pad_0038[8];	//0x0038
-
-		//	
-		virtual void Function0();
-		virtual int GetLevel() = 0;
-		virtual void SetLevel(int lvl);
-		virtual float GetSpAtkTime() = 0;
-		virtual void SetSpAtkTime(float time);
-		virtual float GetSpAtkTimeMax() = 0;
-		virtual void SetSpAtkTimeMax(float maxTime);
-		virtual float GetAttackRate() = 0;
-		virtual void SetAttackRate(float rate);
-		virtual int GetHP() = 0;
-		virtual void SetHP(int hp);
-		virtual int GetHPMax() = 0;
-		virtual void SetHPMax(int mp);
-		virtual void Function13();
-		virtual void Function14();
-		virtual void Function15();
-		virtual void Function16();
-		virtual void Function17();
-		virtual void Function18();
-		virtual int GetAttack() = 0;
-		virtual void SetAttack(int atk);
-		virtual int GetMagicAtk() = 0;
-		virtual void SetMagicAtk(int atk);
-		virtual int GetDefense() = 0;
-		virtual void SetDefense(int def);
-		virtual int GetMagicDef() = 0;
-		virtual void SetMagicDef(int def);
-		virtual void Function27();
-		virtual void Function28();
-		virtual void Function29();
-		virtual void Function30();
-		virtual void Function31();
-		virtual void Function32();
-		virtual void Function33();
-		virtual void Function34();
-		virtual void Function35();
-		virtual void Function36();
-
-		bool IsValid();
-	};	//Size: 0x0040
-	
-	bool ATarget::IsValid()
-	{
-		static __int64 TargetEntity_vfTable = g_GameData->g_GameBaseAddr + 0x4B3E5E8;
-		__int64 vfTable_addr = *(__int64*)((__int64)&this->Level - 0x8);
-		if (!vfTable_addr || vfTable_addr != TargetEntity_vfTable)
-			return false;
-
-		return true;
-	}
-
-	__int64 Hooking::ATargetEntity_GetHP_hook(__int64 p)
+	//	
+	__int64 AGame::Hooks::ATargetEntity_GetHP_hook(__int64 p)
 	{
 		static bool bEditing{ false };	//	used so can call this function without stack overflow
 		auto pTarget = reinterpret_cast<ATarget*>(p);
@@ -649,7 +624,7 @@ namespace FF7Remake
 			bEditing = true;
 
 			//	target health is restored ( note: target can still die if damage taken is > maxhp )
-			if (g_GameData->bNullTargetDmg)
+			if (AGame::bNullTargetDmg)
 			{
 				auto hp_max = pTarget->GetHPMax();
 				auto hp = pTarget->GetHP();	//	recursion
@@ -658,34 +633,34 @@ namespace FF7Remake
 			}
 
 			//	target is killed
-			if (g_GameData->bKillTarget)
+			if (AGame::bKillTarget)
 			{
 				//	if (g_GameData->bMaxTargetLevel)
 				//	{
 				if (pTarget->GetLevel() != 9999)
 					pTarget->SetLevel(9999);
-					//	g_GameData->bMaxTargetLevel = false;
-				//	}
+				//	g_GameData->bMaxTargetLevel = false;
+			//	}
 
 				if (pTarget->GetHP() > 0)
 					pTarget->SetHP(0);
 
-				g_GameData->bKillTarget = false;
+				AGame::bKillTarget = false;
 			}
 
 			//	target wont attack player
-			if (g_GameData->bNoTargetAttack)
+			if (AGame::bNoTargetAttack)
 			{
 				if (pTarget->GetAttackRate() > 0.0f)
 					pTarget->SetAttackRate(0.0f);
 				if (pTarget->GetSpAtkTimeMax() > 0.0f)
 					pTarget->SetSpAtkTimeMax(0.0f);
-				
+
 				//	g_GameData->bNoTargetAttack = false;
 			}
 
 			//	prevents targets from attacking , sets max level and sets health to 0
-			if (g_GameData->bXpFarn)
+			if (AGame::bXpFarn)
 			{
 				if (pTarget->GetLevel() != 9999)
 					pTarget->SetLevel(9999);
@@ -698,7 +673,7 @@ namespace FF7Remake
 
 				if (pTarget->GetDefense() > 0)
 					pTarget->SetDefense(0);
-				
+
 				if (pTarget->GetMagicDef() > 0)
 					pTarget->SetMagicDef(0);
 
@@ -708,24 +683,248 @@ namespace FF7Remake
 
 			bEditing = false;
 		}
-		return g_Hooking->ATargetEntity_GetHP_stub(p);
+		return ATargetEntity_GetHP_stub(p);
 	}
 
-	struct ATargetStagger
+	//	
+	__int64 AGame::Hooks::ATargetEntity_GetStaggerAmount_hook(__int64 a1)
 	{
-		unsigned char pad_0000[0xCF8];	//	0x0000
-		float Stagger;					//	0x0CF8
-		float StaggerMax;				//	0x0CFC
-	};
-	__int64 Hooking::ATargetEntity_GetStaggerAmount_hook(__int64 a1)
-	{
-		if (g_GameData->bXpFarn || g_GameData->bTargetAlwaysStagger)
+		if (AGame::bXpFarn || AGame::bTargetAlwaysStagger)
 		{
 			auto pEnt = reinterpret_cast<ATargetStagger*>(a1);
 			if (pEnt)
 				pEnt->Stagger = pEnt->StaggerMax;
 		}
 
-		return g_Hooking->ATargetEntity_GetStaggerAmount_stub(a1);
+		return ATargetEntity_GetStaggerAmount_stub(a1);
 	}
+
+
+#pragma endregion
+
+	//
+	//	AItemSlot
+	//
+#pragma region	//	AITEMSLOT
+
+	bool AItemSlot::IsValidIndex()
+	{
+		if (flag > 0)
+			return false;
+
+		if (flag2 > 0)
+			return false;
+
+		return isAvailable;
+	}
+
+#pragma endregion
+
+	//
+	//	APlayerStats
+	//
+#pragma region	//	APLAYERSTATE
+
+	void APlayerStats::RefillHP() { HP = MaxHP; }
+
+	void APlayerStats::RefillMana() { MP = MaxMP; }
+
+	void APlayerStats::SetMaxLimit() { LimitBreak = 1000.f; }
+
+	void APlayerStats::SetMaxATB() { ATB = 2000.f; }
+
+#pragma endregion
+
+	//
+	//	ACloudState
+	//
+#pragma region	//	ACLOUDSTATE
+
+	struct AGameState* ACloudState::GetGameState()
+	{
+		return reinterpret_cast<AGameState*>(reinterpret_cast<__int64>(pad_0000) - 0x880);
+	}
+
+#pragma endregion
+
+	//
+	//	AGAMESTATE
+	//
+#pragma region	//	AGAMESTATE
+
+	struct APlayerStats AGameState::GetPlayerStats(int index) { return this->mPartyStats[index]; }
+
+	void AGameState::SetPlayerStats(int index, const APlayerStats newStats) { this->mPartyStats[index] = newStats; }
+
+	struct APlayerStats AGameState::GetCloudStats() { return this->mCloudState.mStats; }
+
+	void AGameState::SetCloudStats(const APlayerStats newState) { this->mCloudState.mStats = newState; }
+
+	struct APlayerAttributes AGameState::GetPlayerAttributes(int index) { return this->mPartyAttributes[index]; }
+
+	void AGameState::SetPlayerAttributes(int index, const APlayerAttributes newAttributes) { this->mPartyAttributes[index] = newAttributes; }
+
+	struct APlayerAttributes AGameState::GetCloudAttributes() { return GetPlayerAttributes(0); }											//	return this->mPamPartyAttributes[0];
+
+	void AGameState::SetCloudAttributes(const APlayerAttributes newAttributes) { SetPlayerAttributes(0, newAttributes); }	//	this->mPartyAttributes[0] = newAttributes; }
+
+#pragma endregion
+
+	//
+	//	AGameBase
+	//
+#pragma region	//	AGAMEBASE
+
+	bool AGameBase::Valid() { return this->mMatchState_0 <= 2; }
+
+	class AGameState* AGameBase::GetGameState()
+	{
+		AGameState* result{ nullptr };
+
+		if (this->Valid())
+			result = this->pGameState;
+
+		return result;
+	}
+
+#pragma endregion
+
+	//
+	//	AScene
+	//
+#pragma region	//	ASCENE
+
+	void AScene::SetPauseState(bool newState) { this->bPause = newState; }
+
+	bool AScene::GetPauseState() { return this->bPause; }
+
+	void AScene::SetTimeScale(float newScalar) { this->TimeScale = newScalar; }
+
+	float AScene::GetTimeScale() { return this->TimeScale; }
+
+#pragma endregion
+
+	//
+	//	OnUseItem
+	//
+#pragma region	//	ON USE ITEM
+
+	AGameState* OnUseItem::GetGameState() { return pCloudState->GetGameState(); }
+	
+	AInventory* OnUseItem::GetInventory() { return reinterpret_cast<AInventory*>(*pInventory); }
+
+	AItemSlot* OnUseItem::GetCurrentItemSlot() { return GetItemSlot(); }
+
+	AItemSlot* OnUseItem::GetItemSlot(int index)
+	{
+		/*
+			AItemSlot* result{ nullptr };
+			auto p = GetInventory();
+			auto i = index * sizeof(AItemSlot);
+			auto z = reinterpret_cast<long long>(p) + i;
+			auto inv = reinterpret_cast<AInventory*>(z);
+			AItemSlot* result = &inv->Item;
+		*/
+		return &reinterpret_cast<AInventory*>(reinterpret_cast<__int64>(GetInventory()) + index * sizeof(AItemSlot))->Item;
+	}
+
+	int OnUseItem::GetCurrentItemCount()
+	{
+		int result{ 0 };
+
+		auto inventory = GetInventory();
+
+		if (inventory)
+			result = inventory->Item.count;
+
+		return result;
+	}
+
+#pragma endregion
+
+	//
+	//	OnSetHealth
+	//
+#pragma region	//	ON SET HEALTH
+
+	OnSetHealth::OnSetHealth() {}
+	
+	OnSetHealth::OnSetHealth(int i, int hp)
+	{
+		bool isTakeDmg{ false };
+		bool isHealing{ false };
+		AGameBase* pGame = AGame::gGameBase;
+		AGameState* pGameS = pGame->GetGameState();
+		APlayerStats player_stats = pGameS->GetPlayerStats(i - 1); //	due to the way structures are spaced cloud is not included in the player stats array. subtracting 1 resolves this issue.
+		int diff = player_stats.HP - hp;
+		if (diff < 0)	//	if <0 player is healing
+		{
+			isHealing = true;
+			diff = hp - player_stats.HP;
+		}
+		else if (diff > 0)
+			isTakeDmg = true;
+
+		index = i;
+		mNewHP = hp;
+		mDiff = diff;
+		bHealing = isHealing;
+		bTakingDmg = isTakeDmg;
+		mPlayerStats = player_stats;
+		pGameState = pGameS;
+	}
+
+#pragma endregion
+
+	//
+	//	OnSetMana
+	//
+#pragma region	//	ON SET MANA
+
+	OnSetMana::OnSetMana() {}
+
+	OnSetMana::OnSetMana(int i, int mp)
+	{
+		bool isSubMana{ false };
+		bool isHealing{ false };
+		AGameBase* pGame = AGame::gGameBase;
+		AGameState* pGameS = pGame->GetGameState();
+		APlayerStats player_stats = pGameS->GetPlayerStats(i - 1); //	due to the way structures are spaced cloud is not included in the player stats array. subtracting 1 resolves this issue.
+		int diff = player_stats.MP - mp;
+		if (diff < 0)	//	if <0 player is healing mana
+		{
+			isHealing = true;
+			diff = mp - player_stats.MP;
+		}
+		else if (diff > 0)
+			isSubMana = true;
+
+		index = i;
+		mNewMP = mp;
+		mDiff = diff;
+		bHealing = isHealing;
+		bUsingMana = isSubMana;
+		mPlayerStats = player_stats;
+		pGameState = pGameS;
+	}
+
+#pragma endregion
+
+	//
+	//	ATarget
+	//
+#pragma region	//	ATARGET
+
+	bool ATarget::IsValid()
+	{
+		static __int64 TargetEntity_vfTable = g_Engine->g_GameBaseAddr + 0x4B3E5E8;
+		__int64 vfTable_addr = *(__int64*)((__int64)&this->Level - 0x8);
+		if (!vfTable_addr || vfTable_addr != TargetEntity_vfTable)
+			return false;
+
+		return true;
+	}
+
+#pragma endregion
+
 }
